+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.8.3 UNRELEASED
 
+### Added
+
+- Make tuning trainer kwargs overwritable (#300)
+- Allow adding categories to NaNEncoder (#303)
+
 ### Fixed
 
 - Underlying data is copied if modified. Original data is not modified inplace (#263)
@@ -10,6 +15,7 @@
 ### Contributors
 
 - jdb78
+- kigawas
 
 ## v0.8.2 Fix for output transformer (12/01/2021)